python/update_api/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
flask